Removed Gather transformation
