+++ src/plugins/intel_gna/src/gna_graph_patterns.hpp
@@ -12,9 +12,12 @@
 #include "layers/gna_layer_info.hpp"
 #include "log/debug.hpp"
 #include "ops/util/util.hpp"
+#include "pre_post_process/transposition_info.hpp"
 
 namespace ov {
 namespace intel_gna {
+
+using namespace ::pre_post_process;
 
 /**
  * @brief checks if it's a reshape from 4d to 3d tensor
