+++ src/plugins/intel_gna/src/memory/gna_memory_state.cpp
@@ -4,10 +4,10 @@
 
 #include "gna_memory_state.hpp"
 
+#include "pre_post_process/preprocessing.hpp"
 #include "frontend/quantized_layer_params.hpp"
 #include "gna_graph_tools.hpp"
 #include "ie_layouts.h"
-#include "preprocessing.hpp"
 
 namespace ov {
 namespace intel_gna {
@@ -70,7 +70,7 @@ void GNAVariableState::SetState(const InferenceEngine::Blob::Ptr& newState) {
             auto quantized =
                 InferenceEngine::getInjectedData<ov::intel_gna::frontend::QuantizedLayerParams>(state->getInput());
             auto scale_factor = quantized != nullptr ? quantized->_dst_quant.GetScale() : state->scale_factor;
-            ConvertToInt16(static_cast<int16_t*>(state->gna_ptr),
+            pre_post_process::ConvertToInt16(static_cast<int16_t*>(state->gna_ptr),
                            newState->buffer().as<float*>(),
                            1,
                            data_elements,
