+++ src/plugins/intel_gna/src/optimizer/gna_pass_manager.cpp
@@ -43,11 +43,13 @@
 #include "layers/gna_layer_info.hpp"
 #include "log/debug.hpp"
 #include "log/log.hpp"
+#include "pre_post_process/transposition_info.hpp"
 
 using namespace InferenceEngine;
 using namespace InferenceEngine::details;
 using namespace ov::intel_gna::frontend;
 using namespace ov::intel_gna::common;
+using namespace ov::intel_gna::pre_post_process;
 
 namespace ov {
 namespace intel_gna {
