--- src/plugins/intel_gna/src/transformations/gather_remove.hpp
@@ -1,27 +1,0 @@
-// Copyright (C) 2022 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-#pragma once
-
-#include <ngraph/pass/graph_rewrite.hpp>
-
-#include "gna_data_types.hpp"
-
-namespace ov {
-namespace intel_gna {
-namespace pass {
-
-class GatherRemove : public ngraph::pass::FunctionPass {
-public:
-    NGRAPH_RTTI_DECLARATION;
-    GatherRemove(ov::intel_gna::SubgraphCPUMap* subgraph_cpu_map = nullptr) : m_subgraph_cpu_map(subgraph_cpu_map) {}
-    bool run_on_model(const std::shared_ptr<ngraph::Function>& f) override;
-
-private:
-    ov::intel_gna::SubgraphCPUMap* m_subgraph_cpu_map;
-};
-
-}  // namespace pass
-}  // namespace intel_gna
-}  // namespace ov
